+++ VDA-StorageNode-Registry/contracts/facets/VDAStorageNodeManagementFacet.sol
@@ -42,7 +42,7 @@ contract VDAStorageNodeManagementFacet is IStorageNodeManagement {
     * @dev Internal function used in the `addNode()` function. Created for stack deep error
     * @param nodeInfo Node information to store
     */
-  function storeNodeInfo(StorageNodeInput memory nodeInfo) internal virtual {
+  function _storeNodeInfo(StorageNodeInput memory nodeInfo) internal virtual {
     LibStorageNode.NodeStorage storage ds = LibStorageNode.nodeStorage();
     {
       uint nodeId = ++ds._nodeIdCounter;
@@ -160,7 +160,7 @@ contract VDAStorageNodeManagementFacet is IStorageNodeManagement {
       ds._stakedTokenAmount[nodeInfo.didAddress] = ds._stakedTokenAmount[nodeInfo.didAddress] + totalAmount;
     }
 
-    storeNodeInfo(nodeInfo);
+    _storeNodeInfo(nodeInfo);
   }
 
   /**
@@ -369,12 +369,12 @@ contract VDAStorageNodeManagementFacet is IStorageNodeManagement {
 
   /**
     * @notice Filter nodes with inputed status
-    * @dev Used for `getNodesByCountry()` and `getNodesByRegion()` functions
+    * @dev Used for `getNodesByCountryCode()` and `getNodesByRegionCode()` functions
     * @param ids ID set
     * @param status Target status of storage nodes
     * @return StorageNode[] Array of active storage nodes
     */
-  function filterStorageNodes(EnumerableSet.UintSet storage ids, LibCommon.EnumStatus status) internal view virtual returns(LibStorageNode.StorageNode[] memory) {
+  function _filterNodesByStatus(EnumerableSet.UintSet storage ids, LibCommon.EnumStatus status) internal view virtual returns(LibStorageNode.StorageNode[] memory) {
     LibStorageNode.NodeStorage storage ds = LibStorageNode.nodeStorage();
 
     uint count = ids.length();
@@ -412,7 +412,7 @@ contract VDAStorageNodeManagementFacet is IStorageNodeManagement {
    * @param ids Array of nodeIds
    * @return StorageNode[] Array of storage node
    */
-  function getAllStorageNodes(EnumerableSet.UintSet storage ids) internal view virtual returns(LibStorageNode.StorageNode[] memory) {
+  function _getNodesById(EnumerableSet.UintSet storage ids) internal view virtual returns(LibStorageNode.StorageNode[] memory) {
     LibStorageNode.NodeStorage storage ds = LibStorageNode.nodeStorage();
     
     uint count = ids.length();
@@ -430,28 +430,60 @@ contract VDAStorageNodeManagementFacet is IStorageNodeManagement {
   /**
     * @dev see { IStorageNodeManagement }
     */
-  function getNodesByCountry(string calldata countryCode) external view virtual override returns(LibStorageNode.StorageNode[] memory) {
-    return getAllStorageNodes(LibStorageNode.nodeStorage()._countryNodeIds[countryCode]);
+  function getNodesByCountryCode(string calldata countryCode) external view virtual override returns(LibStorageNode.StorageNode[] memory) {
+    return _getNodesById(LibStorageNode.nodeStorage()._countryNodeIds[countryCode]);
   }
 
   /**
     * @dev see { IStorageNodeManagement }
     */
-  function getNodesByCountryAndStatus(string calldata countryCode, LibCommon.EnumStatus status) external view returns(LibStorageNode.StorageNode[] memory) {
-    return filterStorageNodes(LibStorageNode.nodeStorage()._countryNodeIds[countryCode], status);
+  function getNodesByCountryCodeAndStatus(string calldata countryCode, LibCommon.EnumStatus status) external view returns(LibStorageNode.StorageNode[] memory) {
+    return _filterNodesByStatus(LibStorageNode.nodeStorage()._countryNodeIds[countryCode], status);
   }
 
   /**
     * @dev see { IStorageNodeManagement }
     */
-  function getNodesByRegion(string calldata regionCode) external view virtual override returns(LibStorageNode.StorageNode[] memory) {
-    return getAllStorageNodes(LibStorageNode.nodeStorage()._regionNodeIds[regionCode]);
+  function getNodesByRegionCode(string calldata regionCode) external view virtual override returns(LibStorageNode.StorageNode[] memory) {
+    return _getNodesById(LibStorageNode.nodeStorage()._regionNodeIds[regionCode]);
   }
 
   /**
     * @dev see { IStorageNodeManagement }
     */
-  function getNodesByRegionAndStatus(string calldata regionCode, LibCommon.EnumStatus status) external view returns(LibStorageNode.StorageNode[] memory) {
-    return filterStorageNodes(LibStorageNode.nodeStorage()._regionNodeIds[regionCode], status);
+  function getNodesByRegionCodeAndStatus(string calldata regionCode, LibCommon.EnumStatus status) external view returns(LibStorageNode.StorageNode[] memory) {
+    return _filterNodesByStatus(LibStorageNode.nodeStorage()._regionNodeIds[regionCode], status);
+  }
+
+  /**
+    * @dev see { IStorageNodeManagement }
+    */
+  function getNodesByStatus(LibCommon.EnumStatus status) external view returns(LibStorageNode.StorageNode[] memory) {
+    LibStorageNode.NodeStorage storage ds = LibStorageNode.nodeStorage();
+    uint count = ds._nodeIdCounter;
+    uint filteredCount;
+
+    for(uint i; i<count;) {
+      if (ds._nodeMap[i+1].status == status) {
+        ++filteredCount;
+      }
+      unchecked {
+        ++i;
+      }
+    }
+
+    LibStorageNode.StorageNode[] memory nodeList = new LibStorageNode.StorageNode[](filteredCount);
+
+    uint nodeId;
+    for (uint i; i < count;) {
+      if (ds._nodeMap[i+1].status == status) {
+        nodeList[nodeId] = ds._nodeMap[i+1];
+        unchecked {
+          ++nodeId;
+        }
+      }
+      unchecked { ++i; }
+    }
+    return nodeList;
   }
 }